requirements.txt
python-dotenv==1.0.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
twilio==8.12.0
pytz==2024.1 